Design Pattern
1. The used Design Pattern

Creational Pattern - Factory Method

2. Revision classes

modified: Bullets.java
path: core/src/mindustry/content

generated: FireBallBulletType.java, FlameBulletType.java, LightningType.java, MeltDownLaserType.java
path: core/src/mindustry/entities/bullet

3. Reason

It's already been made with Factory method pattern I guess but there were some objects that declare in Bullets class(fireball, basicFlame, pyraFlame, meltdownLaser,lightening) . They seemed to violate OCP(Open Closed Priciple) so they had to be fixed. Because, if the bullets I mentioned above need to be modified, we can't avoid the revision in Bullet class. That's reason why I made some classes which define what type of bullet it is. If we want to add other bullets in the type we made, we can reuse the objects.
